fix hcl tests
