+++ go.mod
@@ -10,6 +10,7 @@ require (
 	github.com/hashicorp/go-version v1.3.0
 	github.com/hashicorp/hcl/v2 v2.10.1
 	github.com/hashicorp/terraform v0.15.3
+	github.com/hectane/go-acl v0.0.0-20190604041725-da78bae5fc95
 	github.com/madlambda/spells v0.4.0
 	github.com/posener/complete v1.2.3
 	github.com/willabides/kongplete v0.2.0
